refactor : BattleService @Repository -> @Service, UserService 이름 오타 수정

+++ src/main/java/com/lgm/my_mong/repository/BattleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BattleRepository extends JpaRepository<Battle, Integer> {
+public interface BattleRepository extends JpaRepository<Battle, Long> {
 }
