+++ src/main/java/com/lgm/my_mong/repository/MongRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MongRepository extends JpaRepository<Mong, Integer> {
+public interface MongRepository extends JpaRepository<Mong, Long> {
 }
